+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `ggtibble()` now warns if `outercols` are not used in either the `caption` or
   the `labs` argument (#13).
+* `ggtibble` and `gglist` objects now work with the ggplot2 `%+%` operator (#16)
 
 # ggtibble 1.0.1